Rename to TableFactory and move it to factories package
