+++ flink-libraries/flink-table/src/test/resources/META-INF/services/org.apache.flink.table.factories.TableFactory
@@ -17,6 +17,6 @@
 org.apache.flink.table.sources.TestFixedFormatTableFactory
 org.apache.flink.table.sources.TestWildcardFormatTableSourceFactory
 =======
-org.apache.flink.table.connectors.TestTableSinkFactory
-org.apache.flink.table.connectors.TestTableSourceFactory
->>>>>>> 2fce6f6765... [FLINK-8866] [table] Create unified interfaces to configure and instatiate TableSinks:flink-libraries/flink-table/src/test/resources/META-INF/services/org.apache.flink.table.connectors.DiscoverableTableFactory
+org.apache.flink.table.factories.TestTableSinkFactory
+org.apache.flink.table.factories.TestTableSourceFactory
+>>>>>>> 2fce6f6765... [FLINK-8866] [table] Create unified interfaces to configure and instatiate TableSinks:flink-libraries/flink-table/src/test/resources/META-INF/services/org.apache.flink.table.factories.TableFactory
