+++ flink-libraries/flink-table/src/test/scala/org/apache/flink/table/factories/TableSourceFactoryServiceTest.scala
@@ -16,58 +16,83 @@
  * limitations under the License.
  */
 
-package org.apache.flink.table.connectors
+package org.apache.flink.table.factories
 
 import org.apache.flink.table.api.{NoMatchingTableFactoryException, ValidationException}
-import org.apache.flink.table.descriptors.ConnectorDescriptorValidator._
-import org.apache.flink.table.descriptors.FormatDescriptorValidator._
+import org.apache.flink.table.descriptors.ConnectorDescriptorValidator.{CONNECTOR_PROPERTY_VERSION, CONNECTOR_TYPE}
+import org.apache.flink.table.descriptors.FormatDescriptorValidator.{FORMAT_PROPERTY_VERSION, FORMAT_TYPE}
 import org.apache.flink.table.descriptors.TableDescriptorValidator
-import org.junit.Assert._
+import org.apache.flink.table.sources.TestWildcardFormatTableSourceFactory
+import org.junit.Assert.assertTrue
 import org.junit.Test
 
 import scala.collection.JavaConverters._
 import scala.collection.mutable
 
-class TableSinkFactoryServiceTest {
+class TableSourceFactoryServiceTest {
+
   @Test
   def testValidProperties(): Unit = {
     val props = properties()
-    assertTrue(TableFactoryService.find(classOf[TableSinkFactory[_]], props.toMap) != null)
+    props.put(CONNECTOR_TYPE, "fixed")
+    props.put(FORMAT_TYPE, "test")
+    assertTrue(TableFactoryService.find(classOf[TableSourceFactory[_]], props.toMap) != null)
   }
 
   @Test(expected = classOf[NoMatchingTableFactoryException])
   def testInvalidContext(): Unit = {
     val props = properties()
     props.put(CONNECTOR_TYPE, "FAIL")
-    TableFactoryService.find(classOf[TableSinkFactory[_]], props.toMap)
+    props.put(FORMAT_TYPE, "test")
+    TableFactoryService.find(classOf[TableSourceFactory[_]], props.toMap)
   }
 
   @Test
   def testDifferentContextVersion(): Unit = {
     val props = properties()
+    props.put(CONNECTOR_TYPE, "fixed")
+    props.put(FORMAT_TYPE, "test")
     props.put(CONNECTOR_PROPERTY_VERSION, "2")
     // the table source should still be found
-    assertTrue(TableFactoryService.find(classOf[TableSinkFactory[_]], props.toMap) != null)
+    assertTrue(TableFactoryService.find(classOf[TableSourceFactory[_]], props.toMap) != null)
   }
 
   @Test(expected = classOf[ValidationException])
   def testUnsupportedProperty(): Unit = {
     val props = properties()
+    props.put(CONNECTOR_TYPE, "fixed")
+    props.put(FORMAT_TYPE, "test")
     props.put("format.path_new", "/new/path")
-    TableFactoryService.find(classOf[TableSinkFactory[_]], props.toMap)
+    TableFactoryService.find(classOf[TableSourceFactory[_]], props.toMap)
   }
 
   @Test(expected = classOf[IllegalArgumentException])
   def testFailingFactory(): Unit = {
     val props = properties()
+    props.put(CONNECTOR_TYPE, "fixed")
+    props.put(FORMAT_TYPE, "test")
     props.put("failing", "true")
-    TableFactoryService.find(classOf[TableSinkFactory[_]], props.toMap)
-      .asInstanceOf[TableSinkFactory[_]].createTableSink(props.asJava)
+    TableFactoryService.find(classOf[TableSourceFactory[_]], props.toMap)
+      .asInstanceOf[TableSourceFactory[_]]
+      .createTableSource(props.asJava)
+  }
+
+  @Test
+  def testWildcardFormat(): Unit = {
+    val props = properties()
+    props.put(CONNECTOR_TYPE, "wildcard")
+    props.put(FORMAT_TYPE, "test")
+    props.put("format.type", "not-test")
+    props.put("format.not-test-property", "wildcard-property")
+    val actualTableSource = TableFactoryService.find(classOf[TableSourceFactory[_]], props.toMap)
+    assertTrue(actualTableSource.isInstanceOf[TestWildcardFormatTableSourceFactory])
   }
 
   private def properties(): mutable.Map[String, String] = {
     val properties = mutable.Map[String, String]()
-    properties.put(TableDescriptorValidator.TABLE_TYPE, "sink")
+    properties.put(
+      TableDescriptorValidator.TABLE_TYPE,
+      TableDescriptorValidator.TABLE_TYPE_VALUE_SOURCE)
     properties.put(CONNECTOR_TYPE, "test")
     properties.put(FORMAT_TYPE, "test")
     properties.put(CONNECTOR_PROPERTY_VERSION, "1")
